backend/diagram/requirements.txt
langchain==0.3.7
langgraph==0.2.38
langchain-anthropic==0.2.4
langchain-core==0.3.15
langchain-community==0.3.5
pydantic==2.9.2
python-dotenv==1.0.1
typing-extensions==4.12.2
annotated-types==0.7.0